+++ src/components/dashboard/PageContainer/PageContainer.js
@@ -1,0 +1,12 @@
+/* The main container that wraps the body of all pages */
+
+import React from 'react';
+import { VStack } from '@chakra-ui/react';
+
+const PageContainer = ({ children }) => (
+  <VStack spacing={10} w="100%" h="100%" align="flex-start">
+    {children}
+  </VStack>
+);
+
+export default PageContainer;
