+++ src/pages/Dashboard/Leaderboard.js
@@ -1,9 +1,13 @@
 import React from 'react';
 import { useQuery } from '@apollo/client';
-import { Box, Flex, VStack, Heading, Text, Table, Tr, Td, Th, Thead, Tbody, Spinner, Center } from '@chakra-ui/react';
 import { useSelector } from 'react-redux';
-import { Redirect } from 'react-router-dom';
+import { Flex, Heading, Skeleton, Text, Table, Tr, Td, Th, Thead, Tbody } from '@chakra-ui/react';
+
 import { GET_GLOBAL_LEADERBOARD } from '../../data/gql/team';
+
+import PageHeading from 'components/dashboard/PageHeading/PageHeading';
+
+const headers = ['RANK', 'TEAM NAME', 'AFFILIATION', 'AMOUNT RAISED'];
 
 const composeTableHeader = (header, i) => {
   return (
@@ -60,51 +64,42 @@ const composeTableBody = (row, i) => {
 };
 
 // Base leaderboard page
-const Leaderboard = () => {
+const Leaderboard = ({ team }) => {
   const {
     user: { teamId },
-    team,
   } = useSelector(state => state.auth);
-  const { loading: teamLoading, data: teamData } = team;
 
   const { loading: leaderboardLoading, error, data: leaderboardData } = useQuery(GET_GLOBAL_LEADERBOARD);
-  const headers = ['RANK', 'TEAM NAME', 'AFFILIATION', 'AMOUNT RAISED'];
 
-  return teamLoading || leaderboardLoading ? (
-    <Center h="100%">
-      <Spinner size="xl" />
-    </Center>
-  ) : error ? (
-    `Error! ${error.message}`
-  ) : teamData ? (
-    <Box>
-      <Flex justifyContent="space-between">
-        <VStack alignItems="flex-start">
-          <Heading textTransform="uppercase" as="p" size="subtitle" color="black" opacity="0.5" mb="8px">
-            Team {teamData.teamName}
-          </Heading>
-          <Heading as="h1" size="h1">
-            Leaderboard
-          </Heading>
-        </VStack>
-        <VStack alignItems="flex-end">
-          <Text>YOUR RANK</Text>
-          <Heading as="h1" size="h1">
-            {leaderboardData.getGlobalLeaderboard.findIndex(obj => {
-              return obj.id === teamId;
-            }) + 1}
-          </Heading>
-        </VStack>
+  return (
+    <>
+      <Flex w="100%" justifyContent="space-between">
+        <PageHeading teamName={team.teamName} title="Leaderboard" />
+
+        {leaderboardData && (
+          <Flex direction="column" justify="space-between" align="flex-end">
+            <Text>YOUR RANK</Text>
+            <Heading as="h2" size="h1">
+              {leaderboardData.getGlobalLeaderboard.findIndex(obj => {
+                return obj.id === teamId;
+              }) + 1}
+            </Heading>
+          </Flex>
+        )}
       </Flex>
-      <Table maxHeight="100vh" size="lg">
-        <Thead>
-          <Tr>{headers.map(composeTableHeader)}</Tr>
-        </Thead>
-        <Tbody>{leaderboardData.getGlobalLeaderboard.map(composeTableBody)}</Tbody>
-      </Table>
-    </Box>
-  ) : (
-    <Redirect to="/" />
+      <Skeleton height="400px" isLoaded={!leaderboardLoading} w="100%">
+        {error ? (
+          `Error! ${error.message}`
+        ) : (
+          <Table maxH="100vh" size="lg">
+            <Thead>
+              <Tr>{headers.map(composeTableHeader)}</Tr>
+            </Thead>
+            <Tbody>{leaderboardData?.getGlobalLeaderboard?.map(composeTableBody)}</Tbody>
+          </Table>
+        )}
+      </Skeleton>
+    </>
   );
 };
